Doc-Strings for Tsunamibayes Module
Mostly finalized doc-strings for all the functions in tsunamibayes.
Some ambiguity in the units for some of the functions in utils.py to be clarified later.
I will also seek to identify places in the code where more documentation is necessary to understand the function. (for example, there are some magic numbers present that are hard to follow).